requirements.txt
requests
speechrecognition
gtts
playsound
python-dotenv
nltk
spacy
